backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
python-socketio[asyncio]==5.11.4
starlette==0.46.2
pydantic==2.11.7
pydantic[email]==2.11.7
pymongo[srv]==4.9.1
motor==3.6.0
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
PyPDF2==3.0.1
google-generativeai==0.8.3
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
requests==2.32.4
email-validator==2.1.1